// lib/types.ts
export type NavItem = {
  label: string;
  href: string;
};

export type GalleryImage = {
  src: string;
  alt: string;
  title: string;
  subtitle: string;
};

export type VideoItem = {
  title: string;
  youtubeEmbedUrl: string;
};

export type SocialLink = {
  label: string;
  href: string;
  iconSrc: string;
};

export type ServiceItem = {
  label: string;
};